V2_OLD/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiosqlite
passlib[bcrypt]
mistune
python-frontmatter
httpx
python-slugify
